requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
sqlmodel==0.0.22
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
Pillow==10.4.0
psycopg2-binary==2.9.9